+++ src/zope/sqlalchemy/datamanager.py
@@ -12,9 +12,13 @@
 #
 ##############################################################################
 
+
 from weakref import WeakKeyDictionary
 
+from pkg_resources import parse_version
+
 import transaction as zope_transaction
+from sqlalchemy import __version__ as sqlalchemy_version
 from sqlalchemy.engine.base import Engine
 from sqlalchemy.exc import DBAPIError
 from sqlalchemy.orm.exc import ConcurrentModificationError
@@ -65,6 +69,8 @@ NO_SAVEPOINT_SUPPORT = {"sqlite"}
 _SESSION_STATE = WeakKeyDictionary()  # a mapping of session -> status
 # This is thread safe because you are using scoped sessions
 
+SA_GE_14 = parse_version(sqlalchemy_version) >= parse_version('1.4.0')
+
 
 #
 # The two variants of the DataManager.
@@ -84,22 +90,18 @@ class SessionDataManager(object):
             self, session, status, transaction_manager, keep_session=False):
         self.transaction_manager = transaction_manager
 
-        # Support SQLAlchemy 2.0
-        # https://docs.sqlalchemy.org/en/14/orm/session_api.html?highlight=get_transaction#sqlalchemy.orm.Session.get_transaction
-        transaction = (
-            session.get_transaction()
-            if hasattr(session.transaction, "get_transaction")
-            else session.transaction
-        )
+        if SA_GE_14:
+            root_transaction = session.get_transaction() or session.begin()
+        else:
+            # Support both SQLAlchemy 1.0 and 1.1
+            # https://github.com/zopefoundation/zope.sqlalchemy/issues/15
+            _iterate_parents = (
+                getattr(session.transaction, "_iterate_self_and_parents", None)
+                or session.transaction._iterate_parents
+            )
+            root_transaction = _iterate_parents()[-1]
 
-        # Support both SQLAlchemy 1.0 and 1.1
-        # https://github.com/zopefoundation/zope.sqlalchemy/issues/15
-        _iterate_parents = (
-            getattr(transaction, "_iterate_self_and_parents", None)
-            or transaction._iterate_parents
-        )
-
-        self.tx = _iterate_parents()[-1]
+        self.tx = root_transaction
         self.session = session
         transaction_manager.get().join(self)
         _SESSION_STATE[session] = status
@@ -159,14 +161,6 @@ class SessionDataManager(object):
         subtransactions.
         """
 
-        # Support SQLAlchemy 2.0
-        session = self.session
-        transaction = (
-            session.get_transaction()
-            if hasattr(session.transaction, "get_transaction")
-            else session.transaction
-        )
-
         # ATT: the following check is weak since the savepoint capability
         # of a RDBMS also depends on its version. E.g. Postgres 7.X does not
         # support savepoints but Postgres is whitelisted independent of its
@@ -174,7 +168,7 @@ class SessionDataManager(object):
         # into account (ajung)
         if set(
             engine.url.drivername
-            for engine in transaction._connections.keys()
+            for engine in self.tx._connections.keys()
             if isinstance(engine, Engine)
         ).intersection(NO_SAVEPOINT_SUPPORT):
             raise AttributeError("savepoint")
@@ -317,16 +311,23 @@ class ZopeTransactionEvents(object):
                      self.transaction_manager, self.keep_session)
 
     def before_commit(self, session):
-        # Support SQLAlchemy 2.0
-        transaction = (
-            session.get_transaction()
-            if hasattr(session.transaction, "get_transaction")
-            else session.transaction
+        in_nested_transaction = (
+            session.in_nested_transaction()
+            if SA_GE_14
+            # support sqlalchemy 1.3 and below
+            else session.transaction.nested
         )
         assert (
-            transaction.nested
+            in_nested_transaction
             or self.transaction_manager.get().status == ZopeStatus.COMMITTING
         ), "Transaction must be committed using the transaction manager"
+
+    def do_orm_execute(self, execute_state):
+        dml = any((execute_state.is_update, execute_state.is_insert,
+                   execute_state.is_delete))
+        if execute_state.is_orm_statement and dml:
+            mark_changed(execute_state.session, self.transaction_manager,
+                         self.keep_session)
 
     def mark_changed(self, session):
         """Developer interface to `mark_changed` that preserves the extension's
@@ -380,4 +381,8 @@ def register(
     event.listen(session, "after_bulk_update", ext.after_bulk_update)
     event.listen(session, "after_bulk_delete", ext.after_bulk_delete)
     event.listen(session, "before_commit", ext.before_commit)
+
+    if SA_GE_14:
+        event.listen(session, "do_orm_execute", ext.do_orm_execute)
+
     return ext